+++ src/test/kotlin/com/nylas/resources/MessagesTests.kt
@@ -384,6 +384,7 @@ class MessagesTests {
           sendAt = 1620000000,
           replyToMessageId = "reply-to-message-id",
           trackingOptions = TrackingOptions(label = "label", links = true, opens = true, threadReplies = true),
+          customHeaders = listOf(CustomHeader(name = "header-name", value = "header-value")),
         )
 
       messages.send(grantId, sendMessageRequest)
@@ -430,6 +431,7 @@ class MessagesTests {
               size = 100,
             ),
           ),
+          customHeaders = listOf(CustomHeader(name = "header-name", value = "header-value")),
         )
 
       messages.send(grantId, sendMessageRequest)
@@ -476,6 +478,7 @@ class MessagesTests {
               size = 3 * 1024 * 1024,
             ),
           ),
+          customHeaders = listOf(CustomHeader(name = "header-name", value = "header-value")),
         )
 
       messages.send(grantId, sendMessageRequest)